Simplify for Python 3

+++ dotgrid.py
@@ -80,8 +80,8 @@ def postscript_header(grid_size):
 
 
 def postscript_set_page_size(width, height):
-    width_in_points = int(round(width / point))
-    height_in_points = int(round(height / point))
+    width_in_points = round(width / point)
+    height_in_points = round(height / point)
 
     postscript_set_page_size = f'''\
 << /PageSize [{width_in_points} {height_in_points}] >> setpagedevice
